requirements.txt
numpy==1.13.3
pandas==0.21.0
scikit-learn==0.19.1
scipy==1.0.0
tensorflow-gpu==1.5.0rc0
tensorflow-tensorboard==0.4.0rc3
tqdm==4.19.4